Rename transfer bench file.
